Custom Type: test resolve_params size validation
Add mtr coverage for the DDL-time size checks from the previous commit:
- extension_length_bounds: a resolved persisted_length exceeding
  max_persisted_length is rejected (OVERSIZED_PARAM_TYPE).
- extension_resolve_params_invalid_sizes: a fixed-length persisted_length
  <= 0, a variable-length positive persisted_length, and a non-positive
  max_decode_buffer_length are each rejected (BAD_LEN_PARAM_TYPE,
  VAR_SETS_PERSISTED, BAD_DECODE_BUF).

The four fixture types are added to vsql_test_only.

+++ villagesql/types/type_function.cc
@@ -16,6 +16,7 @@
 
 #include "villagesql/types/type_function.h"
 
+#include <cinttypes>
 #include <cstdlib>
 #include <cstring>
 #include <string>
@@ -133,8 +134,8 @@ bool ResolveParamsFunction::invoke(const std::string &params_str,
   if (result->max_decode_buffer_length <= 0) {
     snprintf(error_msg, VEF_MAX_ERROR_LEN,
              "resolve_params resolved a non-positive max_decode_buffer_length "
-             "(%lld); it must be > 0",
-             static_cast<long long>(result->max_decode_buffer_length));
+             "(%" PRId64 "); it must be > 0",
+             result->max_decode_buffer_length);
     return true;
   }
   if (*endptr == '\0') return false;
